refactor(api): move user-service from shared to iam context
