+++ api/src/identity-access-management/domain/services/user-service.js
@@ -1,7 +1,7 @@
-import { NON_OIDC_IDENTITY_PROVIDERS } from '../../../identity-access-management/domain/constants/identity-providers.js';
-import { AuthenticationMethod } from '../../../identity-access-management/domain/models/AuthenticationMethod.js';
-import { UserToCreate } from '../../../identity-access-management/domain/models/UserToCreate.js';
-import { DomainTransaction } from '../DomainTransaction.js';
+import { DomainTransaction } from '../../../shared/domain/DomainTransaction.js';
+import { NON_OIDC_IDENTITY_PROVIDERS } from '../constants/identity-providers.js';
+import { AuthenticationMethod } from '../models/AuthenticationMethod.js';
+import { UserToCreate } from '../models/UserToCreate.js';
 
 /**
  * @param user
