+++ api/src/identity-access-management/domain/usecases/index.js
@@ -14,7 +14,7 @@ import { mailService } from '../../../shared/domain/services/mail-service.js';
 import * as obfuscationService from '../../../shared/domain/services/obfuscation-service.js';
 import { tokenService } from '../../../shared/domain/services/token-service.js';
 import * as userReconciliationService from '../../../shared/domain/services/user-reconciliation-service.js';
-import * as userService from '../../../shared/domain/services/user-service.js';
+import * as userService from '../services/user-service.js';
 import * as passwordValidator from '../../../shared/domain/validators/password-validator.js';
 import * as userValidator from '../../../shared/domain/validators/user-validator.js';
 import { httpAgent } from '../../../shared/infrastructure/http-agent.js';
