+++ api/src/prescription/organization-learner/domain/usecases/index.js
@@ -1,4 +1,5 @@
 import * as passwordGenerator from '../../../../identity-access-management/domain/services/password-generator.service.js';
+import * as userService from '../../../../identity-access-management/domain/services/user-service.js';
 import * as authenticationMethodRepository from '../../../../identity-access-management/infrastructure/repositories/authentication-method.repository.js';
 import { emailValidationDemandRepository } from '../../../../identity-access-management/infrastructure/repositories/email-validation-demand.repository.js';
 import { lastUserApplicationConnectionsRepository } from '../../../../identity-access-management/infrastructure/repositories/last-user-application-connections.repository.js';
@@ -15,7 +16,6 @@ import { cryptoService } from '../../../../shared/domain/services/crypto-service
 import * as obfuscationService from '../../../../shared/domain/services/obfuscation-service.js';
 import { tokenService } from '../../../../shared/domain/services/token-service.js';
 import * as userReconciliationService from '../../../../shared/domain/services/user-reconciliation-service.js';
-import * as userService from '../../../../shared/domain/services/user-service.js';
 /** TODO
  * Internal API Needed For
  * campaignRepository.getByCode
