+++ api/tests/identity-access-management/integration/domain/services/user-service_test.js
@@ -4,18 +4,18 @@ import { domainBuilder } from '../../../../tooling/domain-builder/domain-builder
 const { pick } = lodash;
 
 import { NON_OIDC_IDENTITY_PROVIDERS } from '../../../../../src/identity-access-management/domain/constants/identity-providers.js';
+import * as userService from '../../../../../src/identity-access-management/domain/services/user-service.js';
 import * as authenticationMethodRepository from '../../../../../src/identity-access-management/infrastructure/repositories/authentication-method.repository.js';
 import * as userRepository from '../../../../../src/identity-access-management/infrastructure/repositories/user.repository.js';
 import { userToCreateRepository } from '../../../../../src/identity-access-management/infrastructure/repositories/user-to-create.repository.js';
 import { getLearnerInfo } from '../../../../../src/prescription/learner-management/infrastructure/repositories/organization-learner-repository.js';
 import * as organizationLearnerRepository from '../../../../../src/prescription/organization-learner/infrastructure/repositories/organization-learner-repository.js';
 import { OrganizationLearnerNotFound } from '../../../../../src/shared/domain/errors.js';
-import * as userService from '../../../../../src/shared/domain/services/user-service.js';
 import { expect } from '../../../../test-helper.js';
 import { databaseBuilder, knex } from '../../../../tooling/databases.js';
 import { catchErr } from '../../../../tooling/test-utils/error.js';
 
-describe('Integration | Domain | Services | user-service', function () {
+describe('Identity Access Management | Integration | Domain | Services | user-service', function () {
   const hashedPassword = 'Abcdef1234';
 
   let user;
